Fix TestShardOrganizer.testShardOrganizerInProgress flakiness
There was a race, the test job was sleeping for only 10ms.
However, since GC pause is possible, sleeping "long enough" should be
measured in seconds. Use a latch for synchronization instead of fixed
sleep time.
